cropssight-backend/requirements.txt
Flask==3.0.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Werkzeug==3.0.1
python-dotenv==1.0.1
google-auth
google-auth-oauthlib
scikit-learn
